Fixed issue on deactivation without enabled toolbar
On activation of the editor you may pass along the option '{disableToolbar: true}'. This will cause the toolbar not to appear. When .deactivate is then called on the object, an error is written to the console on line 500: this.toolbar.style.display = 'none'. The toolbar object is not set. So I simply have added an if(this.toolbar).