``` stringify([space]); ```
### Summary: ###

Return converted editors content to JSON.

### Parameters ###

**space**

Causes the resulting string to be pretty-printed.
The space argument may be used to control spacing in the final string. If it is a number, successive levels in the stringification will each be indented by this many space characters (up to 10). If it is a string, successive levels will indented by this string (or the first ten characters of it).