Adds very basic extension support for Medium Editor.
Allows the user to register extensions that add a button to the editor and can implement various functions. Also includes an example in the README on how to use this to create a simple text highlighter. Also adds a fix so the Editor fails gracefully when the user defines the "buttons" option as empty array.