Using `options.elementsContainer` to calculate ID
Since the container for toolbar and anchor input can be changed to be 
other than `document.body`, in some cases, querying for the number of
`.medium-editor-toolbar` from `document` will not work. This made all
editor instances to have the same ID of 1.

Shadow DOM is an example in which, if `options.elementsContainer` is
set to an element inside it, it will not be available in the Light DOM to
query.

This updated to use `options.elementsContainer` to query for number of
`.medium-editor-toolbar` elements.