Added ability to add forms for extensions
This change passes the editor instance before creating the toolbar so
that we can access properties when creating buttons in extensions. It
also adds functionality to allow extensions to add forms to the toolbar.

It adds the following functions:

* 'initExtensions()' which gets called after the instance is passed.
This allows extensions to create buttons etc.
* `addExtensionForms()` which gets called after the toolbar is created.
This appends any forms from extensions.
* `showForm()` which displays an extension's form.
* `hideForm()` which hides an extension's form

Some other minor changes:

* changed `.medium-editor-toolbar-form-anchor' to more generic
`.medium-editor-toolbar-form` in medium-editor.scss.
* changed `.medium-editor-toolbar-anchor-input' to more generic
`.medium-editor-toolbar-input` in medium-editor.scss.
* Updated all theme .scss files and medium-editor.js to reflect CSS
changes.
* In `setToolbarPosition()` the toolbar is centered horizontally
relative to the editor container when `staticToolbar` is set to `true`.
This looks better when the toolbar has a lot of buttons. Maybe this
should be a property like `centerToolbar`?