Prevent "Argument not optional" error
In IE11 the findAdjacentTextNodeWithContent resulted in the error
"Argument Not Optional". This is because all 4 arguments to
createNodeIterator are required. See
https://msdn.microsoft.com/en-us/library/ie/ff975301%28v=vs.85%29.aspx
which says the 3rd argument should be null and 4th should be false or
true. The 4th argument is ignored by all other browsers since it is no
longer part of the DOM specification. See note on MDN at:
https://developer.mozilla.org/en-US/docs/Web/API/Document.createNodeIterator