expose Util and Selection on the MediumEditor "namespace".
Similar to `statics`, but really is more of a namespace than
a collection of ctors. Some of these functions are superhandy
when doing extensions, and this just exposes them publicly.

Probably drops coverage a ton? Should add regression tests to
the newly-surfaced APIs.

The most useful one that comes to mind is exposing `Util.derives`
so creating an extended AnchorPreview is just:

  function Preview(){ this.base = true; }
  Preview.prototye = { getTemplate(){...} };
  MediumEditor.statics.AnchorPreview = MediumEditor.Util.derives(MediumEditor.statics.AnchorPreview, Preview);