Auto-link: prevent link from disappearing in MSIE
In MSIE, when the cursor is at the end of the paragraph and you hit
enter, an auto-link immediately preceding your cursor would disappear.
This is probably an artifact of how MSIE unlinks an anchor inside a
contenteditable when you press backspace (back delete) before it.

To compensate, the cursor is always moved 'forward' when selection is
imported after an auto-link operation is completed. For instance, if the
cursor is targeted at the end of the text node in this example code:

    <p>Hello: <a href="http://www.example.com">example.com</a></p>

The selection range will be set to (p, 2) instead of to the text node
inside the anchor tag. This prevents MSIE from thinking a backspace has
taken place and removing the anchor tag.