Fix line breaks on Safari
On recent Safari browsers, it's impossible to insert a line break as
Shift + Enter just makes a new paragraph.

Ctrl + Enter works better, creating the <br>, but it also creates a new
paragraph and results in a paragraph inside another.

In addition to checking !event.shiftKey in handleKeyup(), the code
should check for !event.ctrlKey so that execCommand('formatBlock',
false, 'p') won't be called.