Create Util.isKey to check if key is in an event
Instead of always trying to find which key is associated to an event for all browsers, I created a helper in Util to check if the associated key of an event is the one (or multiple) we are looking for.

I follow what is already done by jQuery to detect the key code from an event.

It'll ease the transition when the `.key` will be compatible with all browsers (https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/key)