Use uppercase version of shortcut
We handle shortcut by listening to onKeyDown. There is also onKeyPress but it is only fired when something is inserted in the content. For example, when you hit the shift key, keyPress is never fired because it does not insert something in the content.

onKeyPress & onKeyDown doesn't handle the key code in the same way.
When you hit `b` letter, onKeyDown will receive 66 but onKeyPress will receive 98.
When you hit `B` letter, onKeyDown will receive 66 and onKeyPress will receive 66.
The way we transcript `b` shortcut to a number is by using: 'b'.charCodeAt(0) which is 98

As you might guess, it can't work. Because we will receive 66 when we are waiting for 98 to bold the content;

Using uppercase letter fix the problem because 'B'.charCodeAt(0) is equal to 66.