Auto-link: never import selection to focus on a BR or IMG
In Gecko, importing the selection when the cursor was in an empty
paragraph such as '<p><br /></p>' or '<p><img /></p>' would set the BR
or IMG to be the start and end container of the selected range. The
cursor would get 'stuck' and nothing happened when the user typed.