Linking bugfix: create only one anchor if possible
When selecting text across inline element boundaries, the built-in
browser createLink command will create 2 or more anchor tags. This is a
weird behavior that the user will find strange when they go to delete or
modify one of the anchor tags and find that their changes don't affect
the other adjacent tags that appear to be part of a single link.

This fixes this problem for selections within a single paragraph or
other parent (block-level) element. If the user has selected text across
multiple parent elements, we fall back to the built-in browser behavior
as there is no way to ensure only a single link is created in this
situation.