Cleanup only pasted elements
When you use `cleanPaste` wich a multiline pasted text it will do custom
stuff related to paragraph.

But it will also cleanup *block elements*. The problem is that it cleans
block element of the WHOLE content, not only the pasted one. So I moved
the code that performs this cleanup inside `pasteHTML` where I have a
node so I can cleanup ONLY the pasted content.

Problem appear when using external library that update MediumEditor
content adding custom html, like medium-editor-insert-plugin. It adds
div, figure, etc .. so if the pasted content was multiline, the cleanup
will remove all this tags.