Fix #757, issue where selection restored too large
Fix an issue where the selection import would ignore the embedded
information about empty paragraphs before it and would import a
selection as multiline that should have been contained at the beginning
of a paragraph.

Also restore a range.collapse(true) that was used for an MSIE-specific
fix unrelated to restoring these types of selections.