Don't encode fragment as part of linkValidation
Fix for: https://github.com/yabwe/medium-editor/issues/1237

Previously URIs were split at a question mark to determine path and
query components. This caused a problem when fragments were part of a
link as they were deemed to be part of the query.

This splits a URL at a ? and a # to get both components and does not
encode the fragment. It did feel a little wrong not to encode the
fragment at all as there some characters that should be encoded, however
the encodeUriComponent method is too heavy handed for a fragment.