Updated all methods that set/get the editor content, extensions can change the html that is being set/get.
Moved the extension init before the addElements method to allow
extension hook into the setContent and getContent. Because of the move
the elements are not available in the extension init method, the
extension should subscribe to the add addElement and removeElement
events.

Always enable toolbar when not disabled in options, even when all
elements on page have data-disable-toolbar set, it is possible that
elements are added later that require the toolbar.